+++ app/src/main/java/uz/magnumactive/benefit/remote/models/RegPhoneResp.kt
@@ -6,7 +6,7 @@ data class RegPhoneResp(
     @SerializedName("phone_number") val phone_number: String? = null,
     @SerializedName("result") val result: Boolean? = null,
     @SerializedName("sms_code") val sms_code: Int? = null,
-    @SerializedName("user_id") val user_id: Int? = null,
+    @SerializedName("user_id") val user_id: Long? = null,
     @SerializedName("user_token") val user_token: String? = null,
     @SerializedName("msg") val msg: String? = null
 )
