+++ app/src/main/java/uz/magnumactive/benefit/remote/models/ReqLoginCode.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class ReqLoginCode(
     @SerializedName("device_code") val device_code: String,
-    @SerializedName("user_id") val user_id: Int = AppPrefs.userId,
+    @SerializedName("user_id") val user_id: Long = AppPrefs.userId,
     @SerializedName("user_token") val user_token: String = AppPrefs.userToken!!,
     @SerializedName("phone_number") val phone_number: String = AppPrefs.phoneNumber!!
 )
