+++ app/src/main/java/uz/magnumactive/benefit/remote/models/RespLoginSms.kt
@@ -7,7 +7,7 @@ data class RespLoginSms(
     @SerializedName("first_name") val first_name: String? = null,
     @SerializedName("last_name") val last_name: String? = null,
     @SerializedName("success") val success: Boolean? = null,
-    @SerializedName("user_id") val user_id: Int? = null,
+    @SerializedName("user_id") val user_id: Long? = null,
     @SerializedName("user_token") val user_token: String? = null,
     @SerializedName("data") val data: String? = null
 )
