+++ app/src/main/java/uz/magnumactive/benefit/remote/models/RespUserInfo.kt
@@ -1,8 +1,11 @@
 package uz.magnumactive.benefit.remote.models
 
 import com.google.gson.annotations.SerializedName
-data class RespUserInfo(@SerializedName("user_id") var user_id:Int,
-                        @SerializedName("access_token") var access_token:String,
-                        @SerializedName("user_token") var user_token:String,
-                        @SerializedName("phone_number") var phone_number:String?=null,
-                        @SerializedName("avatar") var avatar:String?=null)
+
+data class RespUserInfo(
+    @SerializedName("user_id") var user_id: Long,
+    @SerializedName("access_token") var access_token: String,
+    @SerializedName("user_token") var user_token: String,
+    @SerializedName("phone_number") var phone_number: String? = null,
+    @SerializedName("avatar") var avatar: String? = null
+)
